fix(config): fix log format specifier for button long press
Added the missing `freertos/task.h` header required for `pdMS_TO_TICKS` and corrected the `ESP_LOGI` format specifier from `%lu` to `%u` with an explicit cast. This resolves clang‑tidy warnings and ensures proper logging of the button long‑press hold time.

+++ main/config_button.c
@@ -3,6 +3,7 @@
 #include "driver/gpio.h"
 #include "esp_log.h"
 #include "freertos/FreeRTOS.h"
+#include "freertos/task.h"
 
 
 #define CONFIG_BUTTON_GPIO                  GPIO_NUM_34
@@ -112,7 +113,7 @@ static void config_button_task(void *arg)
 
         TickType_t hold_time = sys_time - long_press_time_stamp;
         if (pending_long_press && (hold_time >= pdMS_TO_TICKS(config_btn_ctx.long_press_time))) { // Long press event
-            ESP_LOGI(TAG, "Button long press event, hold time: %lu ms", pdTICKS_TO_MS(hold_time));
+            ESP_LOGI(TAG, "Button long press event, hold time: %u ms", (unsigned)pdTICKS_TO_MS(hold_time));
             if (config_btn_ctx.long_press_callback) {
                 config_btn_ctx.long_press_callback(hold_time);
             }
